+++ app/src/main/java/com/mkielar/pwr/Application.kt
@@ -1,0 +1,11 @@
+package com.mkielar.pwr
+
+import android.app.Application
+import org.koin.android.ext.android.startKoin
+
+class Application : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        startKoin(this, listOf(appModule))
+    }
+}
