+++ app/src/main/java/com/mkielar/pwr/email/view/EmailDetailsActivity.kt
@@ -3,15 +3,14 @@ package com.mkielar.pwr.email.view
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.mkielar.pwr.R
-import com.mkielar.pwr.email.viewModel.EmailDetalisDownloader
+import com.mkielar.pwr.email.viewModel.EmailDetailsDownloader
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.main.activity_email_details.*
+import org.koin.android.ext.android.inject
 
 class EmailDetailsActivity : AppCompatActivity() {
-    companion object {
-        const val EMAIL_ID_KEY = "email_id"
-    }
+    private val emailDetailsDownloader: EmailDetailsDownloader by inject()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,12 +19,16 @@ class EmailDetailsActivity : AppCompatActivity() {
         val extras = intent.extras
         if (extras != null) {
             val emailId = extras.getInt(EMAIL_ID_KEY)
-            val disposable = EmailDetalisDownloader(application).fetch(emailId)
+            val disposable = emailDetailsDownloader.fetch(emailId)
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribeOn(Schedulers.newThread())
                 .subscribe { emailDetails ->
                     textView.text = emailDetails.subject
                 }
         }
+    }
+
+    companion object {
+        const val EMAIL_ID_KEY = "email_id"
     }
 }
