+++ app/src/main/java/com/mkielar/pwr/email/viewModel/EmailParser.kt
@@ -1,0 +1,7 @@
+package com.mkielar.pwr.email.viewModel
+
+import com.mkielar.pwr.email.model.Email
+
+interface EmailParser {
+    fun parse(response: String): List<Email>
+}
